# cmd/api/docs/swagger.yaml
definitions:
...
  req.BlockData:
    properties:
      reason:
        type: string
      userid:
        type: integer
    required:
    - reason
    - userid
    type: object
...
  req.CreateAdmin:
    properties:
      email:
        type: string
      password:
        type: string
      user_name:
        type: string
    required:
    - email
    - password
    - user_name
    type: object
  req.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  res.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/admin/adminlogin:
    post:
      consumes:
      - application/json
      description: Logs in an admin user and returns an authentication token
      parameters:
      - description: Admin login details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/req.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admin Login
      tags:
      - Admin
...
  /api/admin/createadmin:
    post:
      consumes:
      - application/json
      description: admin creation
      operationId: CreateAdmin
      parameters:
      - description: New Admin details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/req.CreateAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Create a new admin from admin panel
      tags:
      - Admin
  /api/admin/dashbord/list:
    get:
      consumes:
      - application/json
      description: Admin can access dashboard and view details regarding orders, products,
        etc.
      operationId: admin-dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admin Dashboard
      tags:
      - Admin
  /api/admin/logout:
    post:
      consumes:
      - application/json
      description: Logs out a logged-in admin from the E-commerce web api admin panel
      operationId: admin-logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
      summary: Admin Logout
      tags:
      - Admin
...
  /api/admin/user/block:
    patch:
      consumes:
      - application/json
      description: admin block user access to the store
      parameters:
      - description: User bolocking details
        in: body
        name: blocking_details
        required: true
        schema:
          $ref: '#/definitions/req.BlockData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admin BlockUser
      tags:
      - Admin
  /api/admin/user/unblock/{user_id}:
    patch:
      consumes:
      - application/json
      description: Admins can block users
      operationId: unblock-users
      parameters:
      - description: ID of the user to be blocked
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admin can unbolock a blocked user
      tags:
      - Admin
...
  /api/user/cart/add/{product_items_id}:
    post:
      consumes:
      - application/json
      description: User can add product item to the cart
      operationId: add-to-cart
      parameters:
      - description: product_items_id
        in: path
        name: product_items_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: User add product-item to cart
      tags:
      - Cart
  /api/user/cart/list:
    get:
      consumes:
      - application/json
      description: User can view the cart with amount
      operationId: list-cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: user can view
      tags:
      - Cart
  /api/user/cart/remove/{product_item_id}:
    delete:
      consumes:
      - application/json
      description: User can remove product item from carts
      operationId: remove-from-cart
      parameters:
      - description: product_item_id
        in: path
        name: product_item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: user remove item
      tags:
      - Cart
...
  /user/cart-item/list:
    get:
      consumes:
      - application/json
      description: User can view items in cart
      operationId: list-cart-items
      parameters:
      - description: Page number for pagination
        in: query
        name: count
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: user can view
      tags:
      - Cart
...
swagger: "2.0"